REM-60 refactor: обновление путей импорта хуков и компонентов для улучшения структуры проекта
