+++ frontend/src/hooks/index.ts
@@ -1,1 +1,4 @@
-export { useLogger } from './useLogger';
+export { useLogger } from "./useLogger";
+export { useBulkOperations } from "./useBulkOperations";
+export { useModals } from "./useModals";
+export { useTorrentData } from "./useTorrentData";
