+++ frontend/src/test/components/App/App.adapters.test.tsx
@@ -5,13 +5,13 @@ import App from "../../../App";
 import { TorrentData } from "../../../components/TorrentList";
 import { useTorrentData } from "../../../hooks/useTorrentData";
 import { useModals } from "../../../hooks/useModals";
-import { useFilteredTorrents } from "../../../hooks/useFilteredTorrents";
+import { useFilteredTorrents } from "../../../components/TorrentList/hooks/useFilteredTorrents";
 import { useBulkOperations } from "../../../hooks/useBulkOperations";
 
 // Мокируем хуки
 vi.mock("../../../hooks/useTorrentData");
 vi.mock("../../../hooks/useModals");
-vi.mock("../../../hooks/useFilteredTorrents");
+vi.mock("../../../components/TorrentList/hooks/useFilteredTorrents");
 vi.mock("../../../hooks/useBulkOperations");
 
 // Мокируем компоненты для упрощения тестов
