+++ frontend/src/test/components/App/App.user-interactions.test.tsx
@@ -2,10 +2,10 @@ import React from "react";
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { render, screen, fireEvent } from "@testing-library/react";
 import App from "../../../App";
-import { useFilteredTorrents } from "../../../hooks/useFilteredTorrents";
+import { useFilteredTorrents } from "../../../components/TorrentList/hooks/useFilteredTorrents";
 
 // Мокируем хуки
-vi.mock("../../../hooks/useFilteredTorrents");
+vi.mock("../../../components/TorrentList/hooks/useFilteredTorrents");
 
 // Мокируем компоненты, которые не тестируются в этом файле
 vi.mock("../../../components/Header", () => ({
